some fixes to compile with msvc

+++ include/bcsv/file_header.h
@@ -149,7 +149,7 @@ namespace bcsv {
         static constexpr size_t COLUMN_LENGTH_SIZE = sizeof(uint16_t);          ///< Size per name length: 2 bytes
         
         // Constructors
-        FileHeader(size_t columnCount = 0, uint8_t compressionLevel = 9, uint8_t major = VERSION_MAJOR, uint8_t minor = VERSION_MAJOR, uint8_t patch = VERSION_PATCH);
+        FileHeader(size_t columnCount = 0, size_t compressionLevel = 9, uint8_t major = VERSION_MAJOR, uint8_t minor = VERSION_MAJOR, uint8_t patch = VERSION_PATCH);
         ~FileHeader() = default;
 
         // Version management
@@ -165,7 +165,7 @@ namespace bcsv {
         uint8_t     versionPatch() const                { return header_.versionPatch; }
 
         // Compression management
-        void        setCompressionLevel(size_t level)   { header_.compressionLevel = (level > 9) ? 9 : level; }
+        void        setCompressionLevel(size_t level)   { header_.compressionLevel = (level > 9) ? 9 : static_cast<uint8_t>(level); }
         uint8_t     compressionLevel() const            { return header_.compressionLevel; }
 
         // Flags management
